+++ apps/switchboard/src/index.ts
@@ -1,10 +1,17 @@
+#!/usr/bin/env node
 import {
   KnexAnalyticsStore,
   KnexQueryExecutor,
 } from "@powerhousedao/analytics-engine-knex";
-import { SubgraphManager, getDbClient } from "@powerhousedao/reactor-api";
+import {
+  Subgraph,
+  SubgraphManager,
+  getDbClient,
+} from "@powerhousedao/reactor-api";
 import { PrismaClient } from "@prisma/client";
-import { ReactorBuilder } from "document-drive";
+import { documentModels as atlasDocumentModels } from "@sky-ph/atlas";
+import * as atlasSubgraphs from "@sky-ph/atlas/subgraphs";
+import { ReactorBuilder, driveDocumentModelModule } from "document-drive";
 import RedisCache from "document-drive/cache/redis";
 import { PrismaStorage } from "document-drive/storage/prisma";
 import {
@@ -14,7 +21,6 @@ import {
 import dotenv from "dotenv";
 import express from "express";
 import http from "http";
-import path from "path";
 import { initRedis } from "./clients/redis.js";
 dotenv.config();
 
@@ -39,6 +45,8 @@ const main = async () => {
     const storage = new PrismaStorage(prismaClient);
     const driveServer = new ReactorBuilder([
       documentModelDocumentModelModule,
+      driveDocumentModelModule,
+      ...atlasDocumentModels,
     ] as DocumentModelModule[])
       .withStorage(storage)
       .withCache(redisCache)
@@ -61,15 +69,21 @@ const main = async () => {
     // init router
     await subgraphManager.init();
 
-    // load switchboard-gui
-    app.use(
-      express.static(
-        path.join(
-          __dirname,
-          "../node_modules/@powerhousedao/switchboard-gui/dist",
-        ),
-      ),
-    );
+    for (const [name, module] of Object.entries(
+      atlasSubgraphs as Record<string, Record<string, typeof Subgraph>>,
+    )) {
+      await subgraphManager.registerSubgraph(module[name]);
+    }
+
+    // // load switchboard-gui
+    // app.use(
+    //   express.static(
+    //     path.join(
+    //       __dirname,
+    //       "../node_modules/@powerhousedao/switchboard-gui/dist",
+    //     ),
+    //   ),
+    // );
 
     // start http server
     httpServer.listen({ port: serverPort }, () => {