+++ packages/document-drive/test/read-mode.test.ts
@@ -1,5 +1,16 @@
-import { DocumentModelModule } from "document-model";
+import { createState } from "#drive-document-model/gen/utils";
+import {
+  documentModelDocumentModelModule,
+  DocumentModelModule,
+  PHDocument,
+} from "document-model";
 import { beforeEach, describe, it, vi } from "vitest";
+import createFetchMock from "vitest-fetch-mock";
+import { createDocument } from "../../document-model/dist/src/document-model/gen/utils.js";
+import {
+  DocumentDriveDocument,
+  DocumentDriveState,
+} from "../src/drive-document-model/gen/types.js";
 import {
   ReadDocumentNotFoundError,
   ReadDriveNotFoundError,
@@ -12,23 +23,23 @@ import { DocumentModelNotFoundError } from "../src/server/error.js";
 const fetchMocker = createFetchMock(vi);
 fetchMocker.enableMocks();
 
-const documentModels = Object.values(
-  documentModelsMap,
-) as DocumentModelModule[];
+const documentModels = [
+  documentModelDocumentModelModule,
+] as DocumentModelModule[];
 
-function getDocumentModelModule(id: string) {
-  const documentModel = documentModels.find(
-    (d) => d.documentModelState.id === id,
-  );
+function getDocumentModelModule<TDocument extends PHDocument>(
+  id: string,
+): DocumentModelModule<TDocument> {
+  const documentModel = documentModels.find((d) => d.documentModel.id === id);
   if (!documentModel) {
     throw new Error(`Document model not found for id: ${id}`);
   }
-  return documentModel;
+  return documentModel as unknown as DocumentModelModule<TDocument>;
 }
 
-function buildDrive(state: Partial<DocumentDrive.DocumentDriveState>) {
-  return DocumentDrive.utils.createDocument({
-    state: DocumentDrive.utils.createState({
+function buildDrive(state: Partial<DocumentDriveState>) {
+  return createDocument({
+    state: createState({
       global: state,
     }),
   });
@@ -47,7 +58,7 @@ function buildDocumentResponse(drive: PHDocument) {
   };
 }
 
-function mockAddDrive(url: string, drive: DocumentDrive.DocumentDriveDocument) {
+function mockAddDrive(url: string, drive: DocumentDriveDocument) {
   fetchMocker.mockIf(url, async (req) => {
     const { operationName } = (await req.json()) as {
       operationName: string;